Use custom resource loader for AVPlayer
This allows us to support TLS Client Authentication. Currently only working with direct streams (HLS streams are not loading).

AVPlayer doesn't support providing a URLSessionDelegate directly, however allows providing an AVAssetResourceLoaderDelegate to take control of loading the resource where we can use a URLSession with our own URLSessionDelegate.

Oddly enough, the resource loader only works with schemes the AVPlayer doesn't understand, so we modify the scheme to work around this.